+++ src/QtInterface/Bridge/detail/QtActionVectorConnector.cc
@@ -65,7 +65,8 @@ QtActionVectorConnector::QtActionVectorConnector( QAction* parent, std::vector<C
     for ( size_t j = 0; j < this->state_.size(); j++)
     {
       this->add_connection( this->state_[ j ]->value_changed_signal_.connect(
-        boost::bind( &QtActionVectorConnector::SetActionChecked, qpointer, j, _1, _2 ) ) );
+        boost::bind( &QtActionVectorConnector::SetActionChecked, qpointer, 
+        static_cast< int >( j ), _1, _2 ) ) );
     }
     
     this->add_connection( this->index_->value_changed_signal_.connect(
